+++ framework/post_processors/post_processor.cc
@@ -1,9 +1,7 @@
 #include "framework/post_processors/post_processor.h"
-
-#include "framework/physics/physics_event_publisher.h"
+#include "framework/event_system/physics_event_publisher.h"
 #include "framework/event_system/event_subscriber.h"
 #include "framework/event_system/event.h"
-
 #include "framework/logging/log.h"
 #include "framework/object_factory.h"
 #include <inttypes.h>
